corrected type issue in run interp back related to reduce prod

+++ tfkbnufft/nufft/interp_functions.py
@@ -166,7 +166,7 @@ def run_interp_back(kdat, tm, params):
 
     # initialize output array
     griddat = tf.zeros(
-        shape=(tf.shape(kdat)[0], tf.reduce_prod(dims)),
+        shape=(tf.shape(kdat)[0], tf.cast(tf.reduce_prod(dims), tf.int32)),
         dtype=kdat.dtype,
     )
